+++ modules/mauth-authenticator-scala/src/main/scala/com/mdsol/mauth/scaladsl/utils/ClientPublicKeyProvider.scala
@@ -1,5 +1,7 @@
 package com.mdsol.mauth.scaladsl.utils
 
+import cats.effect.IO
+
 import java.security.PublicKey
 import java.util.UUID
 
@@ -13,4 +15,6 @@ trait ClientPublicKeyProvider {
     * @return Future of { @link PublicKey} registered in MAuth for the application with given appUUID.
     */
   def getPublicKey(appUUID: UUID): Future[Option[PublicKey]]
+
+  def getPublicKeyF(appUUID: UUID): IO[Option[PublicKey]]
 }