+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Swap scalacache-guava to scalacache-caffeine ()
+
 ## [7.0.1] - 2021-09-28
 ### Changed
 Upgraded scala cats lib, and fixed resulting breakages. Also updated sbt plugins.